Magnum: Make gui with Magnum a library -- not header-only
